requirements.txt
python-telegram-bot
requests
apscheduler
psycopg2-binary
sqlalchemy
python-dotenv